Bugfix
Fixed the bug that was causing the Display to freeze up when the
function that set its parent was called effectively preventing the
JFrame from merging with the Display and showing any graphics.

This was being caused by the lack of opengl context on the render
thread, so rendering is now done on the main thread. Similar context
problems exist on dealing with the tick thread but they only show up on
program termination and therefore do not have any effect at runtime.